+++ idea/tests/org/jetbrains/kotlin/idea/refactoring/move/AbstractMoveTest.kt
@@ -19,15 +19,11 @@ package org.jetbrains.kotlin.idea.refactoring.move
 import com.google.gson.JsonObject
 import com.google.gson.JsonParser
 import com.intellij.codeInsight.TargetElementUtilBase
-import com.intellij.ide.highlighter.ModuleFileType
 import com.intellij.openapi.editor.EditorFactory
 import com.intellij.openapi.fileEditor.FileDocumentManager
-import com.intellij.openapi.module.StdModuleTypes
 import com.intellij.openapi.util.io.FileUtil
 import com.intellij.openapi.vfs.VfsUtil
-import com.intellij.openapi.vfs.VfsUtilCore
 import com.intellij.openapi.vfs.VirtualFile
-import com.intellij.openapi.vfs.VirtualFileVisitor
 import com.intellij.psi.*
 import com.intellij.refactoring.BaseRefactoringProcessor.ConflictsInTestsException
 import com.intellij.refactoring.PackageWrapper
@@ -40,7 +36,6 @@ import com.intellij.refactoring.move.moveFilesOrDirectories.MoveFilesOrDirectori
 import com.intellij.refactoring.move.moveInner.MoveInnerProcessor
 import com.intellij.refactoring.move.moveMembers.MockMoveMembersOptions
 import com.intellij.refactoring.move.moveMembers.MoveMembersProcessor
-import com.intellij.testFramework.PsiTestUtil
 import com.intellij.util.ActionRunner
 import org.jetbrains.kotlin.idea.jsonUtils.getNullableString
 import org.jetbrains.kotlin.idea.jsonUtils.getString
@@ -64,8 +59,6 @@ import org.jetbrains.kotlin.test.KotlinTestUtils
 import java.io.File
 
 abstract class AbstractMoveTest : KotlinMultiFileTestCase() {
-    private var isMultiModule = false
-
     protected fun doTest(path: String) {
         val config = JsonParser().parse(FileUtil.loadFile(File(path), true)) as JsonObject
 
@@ -118,27 +111,6 @@ abstract class AbstractMoveTest : KotlinMultiFileTestCase() {
             }
         },
         getTestDirName(true))
-    }
-
-    override fun prepareProject(rootDir: VirtualFile) {
-        if (isMultiModule) {
-            VfsUtilCore.visitChildrenRecursively(
-                    rootDir,
-                    object : VirtualFileVisitor<Any>() {
-                        override fun visitFile(file: VirtualFile): Boolean {
-                            if (!file.isDirectory && file.name.endsWith(ModuleFileType.DOT_DEFAULT_EXTENSION)) {
-                                createModule(File(file.path), StdModuleTypes.JAVA)
-                                return false
-                            }
-
-                            return true
-                        }
-                    }
-            )
-        }
-        else {
-            PsiTestUtil.addSourceContentToRoots(myModule, rootDir)
-        }
     }
 
     protected fun getTestDirName(lowercaseFirstLetter : Boolean) : String {